feat: update event expiry handling to support optional expiry times

+++ amico-core/src/config/core/event_config.rs
@@ -3,5 +3,5 @@ use serde::{Deserialize, Serialize};
 #[derive(Serialize, Deserialize)]
 #[serde(rename_all = "snake_case")]
 pub struct EventConfig {
-    pub expiry_time: u32, // The expiry time of the event in seconds
+    pub expiry_time: i64, // The expiry time of the event in seconds
 }
